move logic to SessionManager.closeSession
