+++ externals/kyuubi-spark-sql-engine/src/main/scala/org/apache/kyuubi/engine/spark/session/SparkSQLSessionManager.scala
@@ -157,23 +157,20 @@ class SparkSQLSessionManager private (name: String, spark: SparkSession)
   }
 
   override def closeSession(sessionHandle: SessionHandle): Unit = {
-    val closeGracefully = conf.get(ENGINE_SPARK_SESSION_CLOSE_GRACEFULLY)
-    if (closeGracefully) {
-      operationManager.waitForOperationsToFinish()
-    }
-    if (!userIsolatedSparkSession) {
-      val session = getSession(sessionHandle)
-      if (session != null) {
-        userIsolatedCacheLock.synchronized {
-          if (userIsolatedCacheCount.containsKey(session.user)) {
-            val (count, _) = userIsolatedCacheCount.get(session.user)
-            userIsolatedCacheCount.put(session.user, (count - 1, System.currentTimeMillis()))
+    try {
+      super.closeSession(sessionHandle)
+
+      if (!userIsolatedSparkSession) {
+        val session = getSession(sessionHandle)
+        if (session != null) {
+          userIsolatedCacheLock.synchronized {
+            if (userIsolatedCacheCount.containsKey(session.user)) {
+              val (count, _) = userIsolatedCacheCount.get(session.user)
+              userIsolatedCacheCount.put(session.user, (count - 1, System.currentTimeMillis()))
+            }
           }
         }
       }
-    }
-    try {
-      super.closeSession(sessionHandle)
     } catch {
       case e: KyuubiSQLException =>
         warn(s"Error closing session ${sessionHandle}", e)
